Index

This page lists the core methods and types of the AdaptiveResonance.jl package. The Methods section lists the public methods for the package that use the modules in Types. Each of these entries link to the docstrings in the Docs section.

ART modules document their internal working parameters and references, while their hyperparameters/options are documented under their corresponding option structs opts_....

Methods

OptimalEstimation.greet

Types

Docs

OptimalEstimation — Module

Main module for OptimalEstimation.jl, a Julia package of optimal state estimation algorithms.

This module exports all of the modules, filters, and utilities used by the OptimalEstimation.jl package.

Exports

greet

OptimalEstimation.greet — Function